LegacyEnvironment#addPluginClasspath: support adding directories
The biggest shortcoming of ImageJ 1.x is a lack of support for adding
*unpacked* plugin .jar files (as you would want when running inside an
IDE, or when running unit tests as part of a Maven build).

This commit adds support for them.

Note: the common workaround was to add the plain .class files into the
plugins/ directory of ImageJ 1.x, but that work-around falls flat on a
couple of accounts: lack of support for plugins.config (and therefore
plugin classes whose names have *no* undercore), lack of support for
dependency .jar files living somewhere completely different, the problem
that using unpacked .class files in plugins/ tests a different code path
than the intended usage, etc.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>