Add test to ensure that batik is 'demoted' to the end of the classpath
ImageJ 1.x' PluginClassLoader operates under the assumption that it will
find the correct class versions no matter how the .jar files are
ordered. Sadly, this is not true: in Fiji, there is a batik.jar with
completely obsolete XML classes, possibly interfering with xalan.jar
which brings up-to-date versions.

To help things, we patch some sane logic into ImageJ 1.x, and this
commit adds the unit test to ensure that this support does not get
broken randomly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>