Avoid anonymous class in LegacyHooks
The LegacyHooks class wants to be shared between class loaders. As such,
it even the LegacyClassLoader needs to be able to find the classes defined
in that class. We cannot do that easily with an anonmous, class, so let's
make the Comparator for fat .jar file names a nested class.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>