LegacyInjector: add a static method to write out a patched .jar file
The code to write out a misc/headless.jar was not quite maintained when
moving the code from ij-legacy into ij1-patcher.

This change reinstates a public API to allow that, and as a bonus, it
also allows to include even the ImageJ 1.x classes that were not patched
as well as the resources needed to run it. In other words, running

	File jarFile = new File("ij.jar");
	boolean headless = false;
	boolean fullJar = true;
	LegacyInjector.writeJar(jarFile, headless, fullJar);

will write out an ij.jar that can be used *without* ij1-patcher.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>