Fix strange FatJarTest failure
The idea of the FatJarTest is to ensure a certain class path entry
order, no matter what the file system order of the files.

To that end, we try very hard to create an *unfavorable* order on the
file system.

After writing xalan.jar, and then batik.jar, we test whether they are
reported in that order by the file system, and rename them when they are
not, in the hope that the order of the directory entries is not affected
by that.

However, under certain circumstances -- e.g. when on an ext4 file system
-- that is not guaranteed when we do too much renaming. Therefore, let's
try to be very careful how we rename the files to minimize the chance of
getting outwitted by the file system.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>