Avoid NullPointerException in the 'duplicate plugins' error
This could only happen if pluginsPath == null, which is now quite
possible. It could not happen earlier, but only due to side effects.
Let's be safe.

This error could be triggered by

	LegacyEnvironment ij1 = new LegacyEnvironment(null, false);
	ij1.runMacro("eval('script', 'new ImageJ(null, 2);');", "");

when plugins.dir was unset and ij1.plugin.dirs pointed to a directory
containing at least one plugin.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>