Fix headless dialog with boolean parameters
Daniel Hornung recognized an issue where the headless dialog would not
work properly. The culprit was that the same label was used for multiple
purposes in the macro options: the file name and the "open all" flag.

In this particular case, it should be non-fatal, though, because the first
use requires a file name while the second, being a flag, requires *no*
value. Example: "open=/tmp/blub.raw" should not match the "open all" flag,
while "open=/tmp/ open" should (yes, it is confusing).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>