LegacyEnvironment: apply legacy patches lazily
So far, ImageJ 1.x is patched completely as soon as the
LegacyEnvironment is instantiated. This is not always desirable,
however: in tests, for example, one will want to disable the
ij.plugin.dirs handling.

So let's prepare the LegacyEnvironment for "lazy" patching, i.e.
waiting with the legacy patches until ImageJ 1.x is accessed. We
can later add methods that exploit the gap between instantiating
the environment and "instantiating" ImageJ 1.x to modify the patches
according to the caller's wishes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>