Provide forward-compatibility with imagej-legacy
The ij-legacy project is just about to migrate into its own source code
repository, being renamed to imagej-legacy and moving into the correct
package: net.imagej.legacy.

As a consequence, we should try to run the initializer in the new package
first, and if it is not found, fall back to the old package name.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>