Add IJ.open hook
It is not sufficient to intercept the openImage method - we want to give
IJ2 a chance to open all resources (e.g. scripts). So we need to
intercept at the IJ.open level in addition to openImage.

Note that IJ.open is void and assumes whatever method it delegates to
handles displaying the opened resource. Thus the hook interceptOpenImage
method returns a boolean to indicate if the provided path was handled or
not.