Make sure that LegacyHooks' fields are retained when installing new ones
We install a patch into ij.IJ._hooks(LegacyHooks) that allows to override
the initially-installed ones. Now that we configure parts of the legacy
environment via fields of the LegacyHooks instance, we have to take care
to copy those settings when installing new hooks.

To make that character of the fields clear, we mark them as private, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>