Refactor the callback handling in LegacyInjector
The previous approach was doomed because it was limited to one
callback, and that callback was required as a parameter at injection
time. However, we are likely to want more callbacks, and we are
likely to know whether we want them to be called *way* earlier
than injection time.

So let's restructure the thing: make a list of callbacks to call
before, and a list to call after, applying the runtime patches.

They are still package-local as they are not considered part of the
public API (we might need to refactor them yet another time before
the API stabilizes).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>