Use the Callback method to disable the ij1.plugin.dirs handling
We added a separate method to the LegacyInjector just to be able to
disable the ij1.plugin.dirs handling. This method does not scale well:
whenever there is a configuration we would want to specify between
instantiating the LegacyEnvironment class and actually using it, we
would have to add a method to both: the LegacyEnvironment and the
LegacyInjector.

Instead, let's use the powerful Callback method to add a patch to the
constructor of the EssentialLegacyHooks class.

Technically, removing the public method from LegacyInjector requires
bumping the major version as per SemVer. However, it was a simple bug
to make this method public instead of package-local (because only the
methods in LegacyEnvironment are really intended for public
consumption). Additionally, we're still in the "0.x" phase and
therefore this developer believes he can get away with claiming that
the removal of the public API that was never intended to be public --
and was in effect only for one minor version anyway -- is actually
a bug fix.

:-)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>