Describe the Callback-based configuration in the README
To increase the bus number of this project, let's keep the README
updated as the internals of the project change. Of course, we describe
the technical details only broadly; the interested user/developer is
supposed to learn about the philosphy from the README (the exact details
can be inferred easily from the source code).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>