Fix JUnit-dependent exception throwing
We had code to log errors when running outside of JUnit without failing,
but to throw errors when running inside JUnit.

However, this code got broken when we realized that we had to have a
specific context class loader while injecting the hooks because we
already loaded the classes right away in that case.

However, we no longer do that: the runtime patches are applied and none
of the code gets called right away. Later, when run through the
LegacyEnvironment, the context class loader is set correctly.

Therefore, let's avoid patching in an extra thread just for the sake of
setting the context class loader.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>