Ask surefire to show full stack traces
In particular when InvocationTargetExceptions are wrapped in other
InvocationTargetExceptions (wrapping other exceptions), the whole package
wrapped in a RuntimeException, surefire tends to like to print only that
RuntimeException's stack trace, followed by the meager information that
there was a cause, namely an InvocationTargetException.

This is not helpful, as we wrap quite a bunch of exceptions in case things
go awry with the runtime patches.

So let's ask the surefire plugin -- nicely -- to please stop trying to be
clever and just give us the full strack traces.

Thankyouverymuch.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>