Verify that stack traces in initializers contain full information
This test was intended to catch the case where tests run via Jenkins
would fail to provide all the information needed to make sense of the
failure.

While this test did not catch that failure (because it was triggered only
by maven-surefire-plugin's insistence to trim stack traces overzealously
by default), it is still a good test to keep.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>