LegacyInjector: intercept 3-param runPlugIn
The 2-string runPlugIn method delegates to the 3-param, but when running
commands from the menu, for example, the 3-param method is called
directly, bypassing the 2-param hook.

As we would like to intercept menu command calls, it makes sense to
intercept the 3-param method instead. The actual command name is not
relevant to the hook right now.