Re-add classes with deprecated package name
This is required for a smooth upgrade path between ij1-patcher versions.
TrackMate, for example, depends on a newer fiji-lib, which in turn
depends on a newer ij1-patcher. But TrackMate does not (yet) depend on
ImageJ2, therefore installing bleeding-edge TrackMate into an existing
Fiji.app/ would completely disrupt the startup.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>