Add a hack to construct SciJava's context robustly
When ImageJ2's LegacyInitializer was found on the class path, it tries
to initialize a SciJava Context. If that fails, no context is installed,
and when that happens inside a Fiji installation, all of Fiji's runtime
patches are skipped.

Clearly, we want to do our best to prevent that from happening.

So when SciJava's Context cannot be created, let's try again with a
Context narrowed down to ImageJ2's DefaultLegacyService.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>