Try to work around ext4 kludginess
Apparently, when having files xalan.jar and batik.jar in a directory in
that native order, renaming a and b can *change* that native order on
the ext4 file system depending on the names they are renamed to.

It *seems* that if we rename batik.jar to xalan.jar instead of
xalax.jar, ext4 somehow does not change their order.

The proper fix, of course, would be not to rely on the natural order,
but what the FatJarTest *actually* tries to verify is that *even if*
the natural order lists the fat .jar files first, we still list them
last. Therefore, relaxing the test would not exactly make certain
that the fat .jar files are listed last...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>