Make HeadlessGenericDialog more robust
We already have a unit test that makes sure that we override all methods
in the HeadlessGenericDialog, in particular those that might have been
added to the GenericDialog class in new ImageJ 1.x versions.

However, this test can only catch things after the fact, after a new
ImageJ version was released (and picked up in Mavenized form by
ImageJA).

Let's be more proactive and add code that inserts dummy methods in the
HeadlessGenericDialog for every unhandled method. This might need to be
switched off for the unit test in question, but it should make the
headless mode a lot more robust.

This fixes https://github.com/imagej/ij1-patcher/issues/13.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>