Add more cautious commitClass(String) method
The commitClass() method is used to make sure that the class loader has
a definition of the given class, possibly after patching said class. If,
however, the class loader in question is the same as the ij1-patcher
lives in, and a Class<?> parameter is required for the commitClass()
method, the class is loaded already, and cannot be patched later
anymore.

Therefore, let's just add a method that takes the class *name*, but keep
the old method for convenience in case we want to commit a class that we
do *not* want to patch anyway.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>