Avoid loading the EssentialLegacyHooks class before patching it
The recent changes to make the LegacyEnvironment configurable involve
optional patches to the EssentialLegacyHooks class. However, if the
class loader to load the patched classes is the very same that hosts
ij1-patcher (i.e. if there is no separate LegacyClassLoader instance),
we really cannot load said class before patching it. So let's be extra
careful.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>