Allow disabling the execution of the initializer class
The initializer class is supposed to be loaded via the PluginClassLoader
first thing after the encapsulated ImageJ 1.x instance has initialized.
It is used at the moment to guarantee a nice startup of Fiji, by letting
the initializer set up a full ImageJ2 context and install the
Fiji-specific extensions.

However, in the near future, we will switch to letting ImageJ2 start
everything up, defaulting to the LegacyUI. So we need a way to disable
the initializer.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>