Allow disabling the PluginClassLoader
This adds the noPluginClassLoader() method to the LegacyEnvironment.
When a legacy environment is configured that way, the patched ImageJ 1.x
will not instantiate a separate class loader to load the plugin. Instead,
it will try to discover the plugins from the same class loader that was
used to load the ImageJ 1.x classes themselves.

As a consequence, it does not make sense to allow users to call
Help>Refresh Menus: there is no way to unload plugin classes, once they
are loaded into the same class loader.

This method was added to support starting up an ImageJ2 context with
a legacy UI that itself starts ImageJ 1.x. If done that way, ImageJ2
and ImageJ 1.x *must* share a class loader, therefore ImageJ 1.x cannot
easily maintain a separate class loader for its plugins (in theory,
we could hardcode which class path elements are part of "core" ImageJ2,
but that would be a quite fragile solution, in particular when one keeps
in mind that personal update sites guarantee certain levels of version
skew to happen).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>