Add a method to check for a running ImageJ 1.x instance gently
Callers who want to run a patched ImageJ 1.x and want to verify that
there is not yet a running ImageJ 1.x instance need to be very careful
*not* to load any ImageJ 1.x classes to perform that check. In
particular, the caller *cannot* test `IJ.getInstance() != null`.

The method `LegacyEnvironment#isImageJ1Initialized(ClassLoader)`
performs such a careful check: if it returns true, there is already
a running instance. If it returns false, we can still apply patches.
If the ImageJ 1.x classes were already loaded, unpatched, this method
throws an exception.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>