Bump parent version to use scijava-common's TestUtils for testing
SciJava common now offers a TestUtils class that has improved versions
of our own createTemporaryDirectory() methods. Let's use them in our
tests, and get rid of our obsolete versions.

To that end, we bump the parent to the pre-emptively deployed
pom-imagej version that already references the ij1-patcher version
we are just about to release.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>