Avoid package-local access in LegacyHooks
The idea of using package-local access was to discourage others from
using a method that is intended only for internal use.

The problem is that package-local access does not only require the
caller to be in the same package. Crucially, it also requires the caller
to be loaded *by the same class loader*.

Since the intended use is, however, to have the LegacyHooks be shared
between class loaders, it is inevitable that the class loaders are
almost *definitely* different.

Let's just make it public, but keep the warning that it is only for
internal use only.

This was an oversight in 2dd84197(Refactor the code to deduce a class
path from a ClassLoader) that was only identified when trying to run
imagej-legacy's legacy service using the new ij1-patcher version.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>