Disallow ImageJ 1.x from parsing plugin .jar files twice
Typically, this would not happen. But we *do* want to be able to extend
the plugin class path, and it is inevitable that we add the same .jar
files that ImageJ 1.x would already add in some circumstances.

Let's just heed Postel's law already and stop ImageJ 1.x from being
clever about 'this cannot happen'.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>