Make sure that bare .class plugins in plugins/**/* are picked up by IJ1
When there is a A_.class file in, say, ImageJ.app/plugins/blub/x/, but the
corresponding class is actually not in the package blub.x, ImageJ 1.x
falls back to trying first x.A_ and then A_. For that to work, the
subdirectories have to be added to the class path.

Needless to say, requiring callers to make sure that the subdirectories
are contained in the class path would be such an inelegant solution that
we rather add rather ugly code to make sure that there is actually a
plugin class loader that contains said directories. But only if really,
really necessary.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>