Headless: Only override the AWT methods in GenericDialog
Methods such as parseDouble() should *not* be stub'bed out to return 0!
The parseDouble() method -- which is another example of conflating UI code
with non-UI logic -- should be the one from GenericDialog.

Unfortunately, the addition of such methods will make it much more
painful, still, to maintain the headless mode (as if it was not hard
enough already). Hopefully ImageJ 1.x will not introduce more offenders
like that. Knock on head.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>