Add a hook to let Macro.getOptions() look at more threads
Unfortunately, Macro.getOptions() is thread local. This assumes that the
number of threads interested in the options is 1 (wrong), and it also
assumes that the same thread should be able only to run one macro (wrong
again!).

Nevertheless, we have to work around the design to fix the case when
scripts are called from within macros and those scripts instantiate
GenericDialogs directly or indirectly that want to use the current macro
options.

This commit introduces the getThreadAncestors() method to the legacy
hooks. By overriding it, the caller can provide a list of thread
ancestors which ImageJ 1.x might have associated with the macro options
in question.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>