Remove silly Macro.getOptions() restriction
ImageJ 1.x insists that macro options can be set anytime, but only
retrieved when the current thread name starts with "Run$_".

Let's not do that.

If there are problems as a consequence, those problems have to be fixed
properly (there could be problems when the options are not correctly
removed, for example, due to missing finally { } blocks). If that is the
case, we will attack them full on instead of papering over by reverting
this commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>