Help ImageJ 1.x avoid an invalid cast
There are so many different code paths to do essentially the same:
create an image and an image window. At least one of those code paths
ends up with a composite image that does not have StackWindow but a
plain ImageWindow.

This happenes e.g. when Fiji's sample image loader uses SCIFIO to load a
16-bit image, as reported by Stefano Ciciliot in
http://fiji.sc/bugzilla/show_bug.cgi?id=813.

ImageJ 1.x is not very careful there and just assumes that every
composite window must have a composite window and casts the window to a
StackWindow to ask whether it has the correct dimensions.

Let's give ImageJ 1.x some help here.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>