Add a way to guard patches against certain ImageJ 1.x versions
The ij1-patcher has to work with a wide range of ImageJ 1.x versions.
The reason is simple: every once in a while, previously working plugins
stop working because ImageJ 1.x development is not *quite* as
backwards-compatible as it could be. In such cases, users are basically
stuck with older versions, leading to the famous 'ImageJ icon gallery'
on the desktop.

Sometimes ImageJ 1.x changes crucially so that completely different
fixes are needed for different versions. To that end, we introduce a
function that lets you verify that the current ImageJ 1.x version is at
least the specified one, for conditional patching.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>