Enable refresh menus with instructions to restart
The previous WARNING message was confusing to users and gave no
indication of what to do to complete installation of a plugin.

Although the Help > Refresh Menus command was disabled to discourage
users from seeing this message, it could still be called from the
command finder and could be called by other plugins (e.g. Plugins >
Install... or Plugins > Install Plugin...).

Let us instead print a helpful message instructing the user to restart
ImageJ.

This change also re-enables the Help > Refresh Menus command, so users
who expect it to work can at least select it and see the instructions to
restart.