Fix erroneous detection of plugins in jars/ folder
We do _not_ want to add all classpath elements to ImageJ1's plugin
classpath. While the SciJava/ImageJ2 approach is to load plugins from
anywhere on the classpath, this is not how ImageJ 1.x was intended to
function. On the contrary: ImageJ1 plugins should only be loaded from
the plugins.dir subtree, and the ImageJ2-only ij1.plugin.dirs subtrees,
as appropriate. It is an error to treat a classpath element outside of
these places as a source of ImageJ 1.x plugins.

This commit also removes a unit test which actively tested for the
undesirable behavior.

See also 908b63d824d56c426d4aefafc5f0242bdf67660d,
which added this behavior in the first place.

Signed-off-by: Curtis Rueden <ctrueden@wisc.edu>