Fix the HeadlessGenericDialog for use with DialogListeners
DialogListeners get the chance to enquire all of the dialog values when
the user hits the OK button, and they actually all need to be enquired,
too, because of the change introduced into ImageJ 1.38u that records the
values in the macro recorder during the final run of the DialogListener
(i.e. when the event is null), as opposed to the usual method to record
the values after the `showDialog()` method returned.

That means that we need to reset the counters after giving the
DialogListener a chance to run, because the regular code after the
`showDialog()` call might want to look at the values *again*.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>