Update ImageJ 1.x from 1.49s to 1.50e
Starting in 1.49u, ImageJ now has a Help > Examples menu with
several submenus of macros illustrating ImageJ's features. Due to
how the Help > Examples submenus are created, it would be rather
involved to runtime patch them such that they are still available
in headless mode. Furthermore, it is highly unlikely that users
will need access to these Example commands from a headless macro.
Hence, we opt to ignore them in the HeadlessCompletenessTest.