EssentialLegacyHooks: skip build directories
When running from an IDE, and when running unit tests, the ij1-patcher
classes may reside in a build directory such as target/classes. Rather
than recursively adding all the package folders to the URLClassLoader in
that case, let's just skip them, because the use case for this recursive
behavior is not relevant in those scenarios.

That is: no one is dropping packageless bare class files into the target
subtree as part of the Maven build. Let's avoid errors like:

    java.lang.IllegalArgumentException: Cannot load class: LegacyImageDisplayService
            at org.scijava.util.ClassUtils.loadClass(ClassUtils.java:217)

Under error-free circumstances, we should not see the above errors.
However, as soon as something goes wrong (e.g. during Context creation),
the ij1-patcher and/or imagej-legacy unit tests will not only have
failing tests, but also spout tons of the above errors due to the
URLClassLoader trying every combination of stripped package prefixes.
With this change, the above junk will no longer be printed.