Adapt to code changes introduced in 1.52t
The getParentFrame() and resetCounters() methods were changed from private to default scope.

As resetCounters() was implemented in HeadlessGenericDialog already, we copy the change to default scope.
For getParentFrame(), we don't require overriding, so ignore this in HeadlessCompletenessTest.

The ij version is pinned to 1.52t until the next release of pom-scijava.