Adapt to code changes introduced in 1.53a
The getParentFrame() is stubified in LegacyHeadless. This wasn't sufficient to let the travis builds pass, so we also try to skip AWT instantiations in ij.gui.GUI, where a new getParentFrame() method was added.

The newly added getLabel() method is private and therefore added to the methods ignored in HeadlessCompletenessTest.

The ij version is pinned to 1.53a until the next release of pom-scijava.