Fix hacker.overrideFieldWrite impl and usage
We only want to replace code if the access is a *write* access.
If it is a read access, we should do nothing.

For the ij.Prefs override, the expression should evaluate to what gets
assigned to the field. The "$_ = " part is unnecessary, and in fact
"$_" is ignored when replacing field write accesses.