Force batch mode always while in headless mode
When creating an ImagePlus and then calling show() on it, while in
headless mode but before any macros are run, a HeadlessException is
thrown. But as soon as a macro is executed, subsequent calls to show()
start working. It turns out this is because running a macro via runMacro
sets Interpreter.batchMode to true, and then batchMode is not set back
to false again afterward (unless the macro is aborted or errors out).

The batchMode flag stays true persistently until one of a number of less
common conditions is met, and so headless mode was reasonably usable.
It works so well that I am going to try just forcing batchMode to stay
true in as many circumstances as possible. Set it to true immediately at
startup. Don't set it back to false if a macro aborts or errors. And
don't ever report it as false from Interpreter.isBatchMode().

This change might break something, but as things stand, the headless
behavior is certainly broken, in that show() does not consistently work.
So let's give it a try for a while and see how it goes!