Open Recent: if String is returned, open it
https://github.com/imagej/imagej-legacy/commit/cc004839c54bf6aa5b4a951f50b953eb4f0383cb
changed the behavior of the DefaultLegacyOpener to return a String
(path) instead of null if the opener doesn't handle the input. This
broke our hacking of the RecentOpener, as we were explicitly looking for
null return values and were short-circuiting if the return value wasn't
an ImagePlus.

Instead, if the result is a String we should just open it.