Remove long-obsolete imagej.patcher package tree
This was left in place for compatibility with very old code that
expects to find the patcher at the old package prefix. But it has
been many years now, and the likelihood of such code still being in
production has greatly diminished. Of course, we still bump the major
version digit to indicate the breakage in backwards API compatibility.